Changes the marshalling of arguments to preserve arguments with self
Addresses issue#15 by inserting the 'self' argument to an expansion's
argument list instead of appending. This solution might be quite
fragile, a more long-term solution might be to add a flag indicating
whether or not a parameter has already been substituted, for example,
or to rewrite the definition of Self class to include this flag.